backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
anthropic==0.28.0
beautifulsoup4==4.12.3
markdownify==0.12.1
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.7.3
aiofiles==23.2.1
python-dotenv==1.0.1
pdfminer.six==20221105
google-generativeai>=0.7.0
